requirements.txt
Flask
opencv-python
numpy
Pillow
ultralytics
facenet_pytorch
torch
